requirements.txt
jax
diffrax
numpyro
optax